chore: migrate benchmarks to Deno.bench
